LayoutTests/http/tests/security/contentSecurityPolicy/media-src-blocked.html

<meta http-equiv="X-WebKit-CSP" content="media-src 'none'">
<video></video>
<script src=../../../media-resources/media-file.js></script>
<script src=../../../media-resources/video-test.js></script>
<p>This test passes if it doesn't alert failure.</p>
<script>
waitForEvent('loadedmetadata', function () {
    alert('FAIL');
} );

addEventListener('load', endTestLater, false);

// Find a supported media file.
var mediaFile = findMediaFile("video", "content/test");
var mimeType = mimeTypeForFile(mediaFile);

video.src = "http://127.0.0.1:8000/resources/load-and-stall.cgi?name=../../../media/" + mediaFile + "&mimeType=" + mimeType + "&stallAt=100000";
</script>
