LayoutTests/http/tests/security/contentSecurityPolicy/media-src-blocked-expected.txt

CONSOLE MESSAGE: Refused to load media from 'http://127.0.0.1:8000/media/video-load-and-stall.cgi?name=../../../media/content/test.mp4&mimeType=video/mp4&stallAt=100000' because of Content-Security-Policy.

END OF TEST
This test passes if it doesn't alert failure.
